+++ test/tasks/report/WeeklyReportNotificationTaskSpec.scala
@@ -5,7 +5,7 @@ import models.company.Address
 import org.specs2.Specification
 import org.specs2.concurrent.ExecutionEnv
 import org.specs2.matcher.FutureMatchers
-import services.MailRetriesService.EmailRequest
+import services.emails.MailRetriesService.EmailRequest
 import utils._
 
 import java.time.OffsetDateTime
